kernel/requirements.txt
langgraph
langchain-openai
pydantic
pydantic-ai
matplotlib
networkx
requests
python-dotenv
